make own function for templates
